+++ aacgmv2/deprecated.py
@@ -20,6 +20,7 @@ Laundal, K. M. and A. D. Richmond (2016), Magnetic Coordinate Systems, Space
 from __future__ import division, absolute_import, unicode_literals
 import numpy as np
 import logbook as logging
+import warnings
 
 def convert(lat, lon, alt, date=None, a2g=False, trace=False, allowtrace=False,
             badidea=False, geocentric=False):
@@ -55,6 +56,10 @@ def convert(lat, lon, alt, date=None, a2g=False, trace=False, allowtrace=False,
         Output longitude in degrees E
     """
     import aacgmv2
+
+    dstr = "Deprecated routine, may be removed in future versions.  Recommend "
+    dstr += "using convert_latlon or convert_latlon_arr"
+    warnings.warn(dstr, category=FutureWarning)
 
     if(np.array(alt).max() > 2000 and not trace and not allowtrace and
        not badidea):
@@ -109,6 +114,10 @@ def subsol(year, doy, utime):
     After Fortran code by A. D. Richmond, NCAR. Translated from IDL
     by K. Laundal.
     """
+    dstr = "Deprecated routine, may be removed in future versions"
+    warnings.warn(dstr, category=FutureWarning)
+
+    
     yr2 = year - 2000
 
     if year >= 2101:
@@ -177,6 +186,10 @@ def gc2gd_lat(gc_lat):
     gd_lat : (same as input)
         Geodetic latitude in degrees N
     """
+    dstr = "Deprecated routine, may be removed in future versions"
+    warnings.warn(dstr, category=FutureWarning)
+
+    
     wgs84_e2 = 0.006694379990141317 - 1.0
     return np.rad2deg(-np.arctan(np.tan(np.deg2rad(gc_lat)) / wgs84_e2))
 
@@ -203,6 +216,10 @@ def igrf_dipole_axis(date):
     """
     import datetime as dt
     import aacgmv2
+
+    dstr = "Deprecated routine, may be removed in future versions"
+    warnings.warn(dstr, category=FutureWarning)
+
 
     # get time in years, as float:
     year = date.year